Extract mapst base class from arrayst
Separate map-theoretic reasoning (index tracking, equality tracking,
Ackermann constraints, read-over-weakeq, extensionality, CDCL(T)
propagation, constraint counting) from array-specific encoding
(with/if/of/comprehension constraints, lazy selects, 2D inlining).

Inheritance chain: arrayst -> mapst -> equalityt -> prop_conv_solvert
